refactor to allow pulling readings from stdin for testing
The overall aim is to create a FakeSensor that consumes readings
from stdin at a rate of 1 per second (total=2 per second, as we
have one for temperature and one for weight). You can then do:

cat teabot/tests/fixtures/empty_then_pot.input \
    | TEABOT_BASE_URL="https://teabot-alsuren.herokuapp.com/" \
        python teabot/worker.py --fake

and it will post a teapot for you.

I had to refactor is_rising_or_constant(), because it pulling
a bunch of readings off in a tight loop when actually it would be
better to look in recent_readings. It will now only block and poll in
a loop if there are less than 10 elements in recent_readings.

I also pulled is_rising_or_constant() out into BaseSensor, so that
FakeSensor could get it for free.