make state machine graph render more nicely
dot will always put edge labels to the right of the line, and does a good job of
avoiding overlapping labels, but doesn't give them any padding by default, so I
added some.

I also want FULL_TEAPOT at the bottom, of the file, so I sort its entries to the
bottom of the dotfile.