debug: I'm actually quite curious about the exceptions Aaron mentions
Apparently there are 'None of the expected conditions were satisfied'
exceptions floating around. Hopefully I will be able to catch them
with dump_sensors().

I expect that it's a simple matter of not calibrating the weight
of the teapot correctly or something silly, but it would be good
to know for sure.